# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ProductCreateRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      dimensions:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      sort_order:
        type: integer
      stock:
        minimum: 0
        type: integer
      weight:
        type: number
    required:
    - name
    - price
    type: object
  models.ProductListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.ProductResponse'
        type: array
      total:
        type: integer
    type: object
  models.ProductResponse:
    properties:
      category_id:
        type: integer
      category_slug:
        type: string
      created_at:
        type: string
      description:
        type: string
      dimensions:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      sort_order:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
      weight:
        type: number
    type: object
  models.ProductUpdateRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      dimensions:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      sort_order:
        type: integer
      stock:
        minimum: 0
        type: integer
      weight:
        type: number
    type: object
...
host: 45.12.229.112:8080
info:
  contact: {}
  description: REST API для управления продуктами с Redis кэшированием
  title: Products API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Возвращает список продуктов с пагинацией и фильтрацией
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        name: limit
        type: integer
      - description: Фильтр по ID категории
        in: query
        name: category_id
        type: string
      - description: Поиск по названию
        in: query
        name: search
        type: string
      - description: Минимальная цена
        in: query
        name: min_price
        type: number
      - description: Максимальная цена
        in: query
        name: max_price
        type: number
      - default: created_at
        description: Сортировка (name, price, created_at)
        in: query
        name: sort
        type: string
      - default: desc
        description: Порядок сортировки (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список продуктов
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Создает новый продукт (требует роль admin)
      parameters:
      - description: Данные продукта
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание продукта
      tags:
      - products
  /products/{id}:
    delete:
      description: Удаляет продукт из системы (требует роль admin)
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление продукта
      tags:
      - products
    get:
      description: Возвращает информацию о продукте по его ID
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение продукта по ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Обновляет информацию о продукте (требует роль admin)
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление продукта
      tags:
      - products
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...